# src/TsutaAI/backend-api/openapi.yaml
openapi: 3.0.3
info:
  title: TsutaAI Backend API
  description: |
    TsutaAI プロジェクト・タスク管理システム バックエンドAPI

    ## 認証について
    ほとんどのエンドポイントはJWT認証が必要です。
    `/auth/login`でトークンを取得し、Authorizationヘッダーに`Bearer {token}`形式で含めてください。

    ## AI機能について
    本システムはGroq APIを使用したAI機能を提供します。
    詳細は[AI利用開示](/docs/AI-USAGE-DISCLOSURE.md)をご確認ください。
  version: 1.0.0
  contact:
    name: TsutaAI Support
  license:
    name: Proprietary
    url: https://github.com/yourusername/TsutaAI/blob/main/LICENSE
servers:
  - url: http://localhost:3001
    description: ローカル開発環境
  - url: https://localhost:3443
    description: ローカル開発環境 (HTTPS)
  - url: https://api.tsutaai.example.com
    description: 本番環境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT トークンによる認証
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: エラーが発生しました
        errorId:
          type: string
          example: ERR-1234567890-abc123
          description: 本番環境でのみ返却されるエラーID
        error:
          type: string
          description: 開発環境でのみ返却されるエラーメッセージ
        stack:
          type: string
          description: 開発環境でのみ返却されるスタックトレース
...
    Project:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 新規Webサイト開発
        description:
          type: string
          example: コーポレートサイトのリニューアル
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
          example: active
        startDate:
          type: string
          format: date
          example: "2025-01-01"
        endDate:
          type: string
          format: date
          example: "2025-12-31"
        ownerId:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /projects:
    get:
      tags:
        - projects
      summary: プロジェクト一覧取得
      description: すべてのプロジェクトを取得（ページネーション対応）
      operationId: getProjects
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [planning, active, on_hold, completed, cancelled]
        - name: search
          in: query
          schema:
            type: string
          description: プロジェクト名で検索
      responses:
        '200':
          description: プロジェクト一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 50
                      totalPages:
                        type: integer
                        example: 3
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - projects
      summary: プロジェクト作成
      description: 新規プロジェクトを作成
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: 新規プロジェクト
                description:
                  type: string
                  example: プロジェクトの説明
                status:
                  type: string
                  enum: [planning, active, on_hold, completed, cancelled]
                  default: planning
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '201':
          description: プロジェクト作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  project:
                    $ref: '#/components/schemas/Project'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{id}:
    get:
      tags:
        - projects
      summary: プロジェクト詳細取得
      description: 指定したIDのプロジェクト詳細を取得
      operationId: getProjectById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: プロジェクト詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  project:
                    $ref: '#/components/schemas/Project'
        '404':
          description: プロジェクトが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - projects
      summary: プロジェクト更新
      description: プロジェクト情報を更新
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [planning, active, on_hold, completed, cancelled]
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  project:
                    $ref: '#/components/schemas/Project'

    delete:
      tags:
        - projects
      summary: プロジェクト削除
      description: プロジェクトを削除（管理者のみ）
      operationId: deleteProject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: プロジェクトを削除しました

  # ========================================
  # タスクエンドポイント
  # ========================================
...